+++ ui-core/src/test-utils/mocking-utils.ts
@@ -9,6 +9,7 @@ import {
   Profile
 } from 'src/types/user'
 import { StudyEnvironmentSurvey } from 'src/types/study'
+import { ParticipantFile } from '../../build-types'
 
 /** simplest survey.  one page, no interactive elements */
 export function generateSurvey(overrideObj?: any): Survey { // eslint-disable-line @typescript-eslint/no-explicit-any
@@ -201,5 +202,16 @@ export const mockProfile = (): Profile => {
       country: 'US',
       postalCode: '02478'
     }
+  }
+}
+
+export const mockParticipantFile = (fileName: string, surveyResponseIds?: string[]): ParticipantFile => {
+  return {
+    id: fileName,
+    fileName,
+    fileType: 'application/pdf',
+    createdAt: 0,
+    lastUpdatedAt: 0,
+    surveyResponseIds: surveyResponseIds || []
   }
 }
