+++ ui-participant/src/test-utils/test-portal-factory.tsx
@@ -1,6 +1,6 @@
 import {
   HtmlPage,
-  LocalSiteContent, ParticipantFile,
+  LocalSiteContent,
   Portal,
   PortalEnvironment,
   PortalEnvironmentConfig,
@@ -131,16 +131,5 @@ export const mockUsePortalEnv = () => {
     reloadPortal: jest.fn(),
     portalEnv: mockPortalEnvironment(),
     localContent: mockLocalSiteContent()
-  }
-}
-
-export const mockParticipantFile = (fileName: string, surveyResponseIds?: string[]): ParticipantFile => {
-  return {
-    id: fileName,
-    fileName,
-    fileType: 'application/pdf',
-    createdAt: 0,
-    lastUpdatedAt: 0,
-    surveyResponseIds: surveyResponseIds || []
   }
 }
